+++ database/migrations/2026_04_10_013229_create_event_rsvps_table.php
@@ -11,9 +11,9 @@ return new class extends Migration
     {
         Schema::create('event_rsvps', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('event_id');
+            $table->unsignedBigInteger('event_id');
             $table->foreign('event_id')->references('id')->on('events')->cascadeOnDelete();
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->cascadeOnDelete();
             $table->enum('status', ['attending', 'maybe', 'declined']);
             $table->text('note')->nullable();
